chore!: move cssdocs to schema-extract
