+++ packages/core/src/index.ts
@@ -17,7 +17,6 @@ export * from './custom-values';
 export * from './state-validators';
 export * from './selector-utils';
 export * from './native-reserved-lists';
-export * from './cssdocs';
 export * from './resolve-namespace-factories';
 export * from './module-resolver';
 export * from './report-diagnostic';
