+++ packages/schema-extract/src/index.ts
@@ -1,2 +1,3 @@
 export * from './main';
+export * from './cssdocs';
 export * from './types';
